[#245] feat: 일정 수정 리팩토링
