+++ lib/router/router_info.dart
@@ -33,9 +33,11 @@ import 'package:tripStory/view/splash/binding/splash_binding.dart';
 import 'package:tripStory/view/splash/views/splash_view.dart';
 import 'package:tripStory/view/trip/bindings/j_plan_add_binding.dart';
 import 'package:tripStory/view/trip/bindings/j_plan_binding.dart';
+import 'package:tripStory/view/trip/bindings/j_plan_edit_binding.dart';
 import 'package:tripStory/view/trip/bindings/location_search_binding.dart';
 import 'package:tripStory/view/trip/bindings/trip_main_binding.dart';
 import 'package:tripStory/view/trip/views/j_plan_create_view.dart';
+import 'package:tripStory/view/trip/views/j_plan_edit_view.dart';
 import 'package:tripStory/view/trip/views/location_search_view.dart';
 import 'package:tripStory/view/trip/views/trip_main_view.dart';
 
@@ -164,6 +166,18 @@ class RouterInfo {
       name: Routes.locationSearch,
       page: () => const LocationSearchView(),
       binding: LocationSearchBinding(),
+    ),
+    GetPage(
+      name: Routes.tripJPlanEdit,
+      page: () {
+        final selectedDate = Get.arguments;
+
+        if (selectedDate is! DateTime) {
+          throw ArgumentError('error');
+        }
+        return JPlanEditView(selectedDate: selectedDate);
+      },
+      binding: JPlanEditBinding(),
     ),
   ];
 }
