+++ lib/util/extension/string_extension.dart
@@ -13,4 +13,16 @@ extension StringExtension on String {
   }
 
   String get formatDeleteSecondTime => substring(0, 5);
+
+  DateTime toTodayTime() {
+    final parts = split(":");
+    if (parts.length < 2) return DateTime.now();
+
+    final now = DateTime.now();
+    final hour = int.tryParse(parts[0]) ?? 0;
+    final minute = int.tryParse(parts[1]) ?? 0;
+    final second = parts.length > 2 ? int.tryParse(parts[2]) ?? 0 : 0;
+
+    return DateTime(now.year, now.month, now.day, hour, minute, second);
+  }
 }
