+++ lib/view/trip/views/j_plan_view.dart
@@ -165,7 +165,7 @@ class _JPlanViewState extends State<JPlanView> {
                                                     items: [
                                                       PopupMenuAction(
                                                         title: "일정 수정",
-                                                        onTap: () {},
+                                                        onTap: () => controller.onEditPlanPressed(),
                                                         iconPath: IconConstants.pencil,
                                                       ),
                                                       PopupMenuAction(
